Prepare for other scanning devices
