+++ BrillouinAcquisition/NIDAQ.h
@@ -1,0 +1,29 @@
+#ifndef NIDAQ_H
+#define NIDAQ_H
+
+#include <QSerialPort>
+#include <vector>
+#include "NIDAQmx.h"
+#include "scancontrol.h"
+
+class NIDAQ: public ScanControl {
+	Q_OBJECT
+
+public:
+	NIDAQ() noexcept;
+	~NIDAQ();
+
+	void setPosition(std::vector<double> position);
+	void setPositionRelative(std::vector<double> distance);
+	std::vector<double> getPosition();
+
+public slots:
+	void init();
+	bool connectDevice();
+	bool disconnectDevice();
+	void setElement(ScanControl::DEVICE_ELEMENT element, int position);
+	void setElements(ScanControl::SCAN_PRESET preset);
+	void getElements();
+};
+
+#endif // NIDAQMX_H
