--- BrillouinAcquisition/thread.cpp
@@ -1,7 +1,0 @@
-#include "stdafx.h"
-#include "thread.h"
-
-void Thread::startWorker(QObject *worker) {
-	start();
-	worker->moveToThread(this);
-}
